refactor: mentoring폴더명 home으로 변경 #82
- pages 폴더 제거
- home폴더 하위에 바로 page파일 위치
