+++ frontend/src/App.tsx
@@ -3,8 +3,8 @@ import { createBrowserRouter, RouterProvider } from 'react-router-dom';
 import MobileLayout from './common/components/MobileLayout/MobileLayout';
 import MentoringBooking from './pages/booking/pages/MentoringBooking';
 import { PAGE_URL } from './pages/common/constants/url';
-import Detail from './pages/mentoring/pages/Detail';
-import Home from './pages/mentoring/pages/Home';
+import Detail from './pages/detail/Detail';
+import Home from './pages/home/Home';
 
 const router = createBrowserRouter([
   { path: PAGE_URL.HOME, element: <Home /> },
